\name{InvenireSRNA-package}
\alias{InvenireSRNA-package}
\docType{package}
\title{
An R package which specializes in predicting CsrA regulating sRNAs and also provides the user with the ability to learn a similar classification model for different classes of sRNAs.
}
\description{
Small RNAs (sRNAs) constitute an important class of post-transcriptional regulators that control critical cellular processes in bacteria. While recent research has led to a dramatic increase in the discovery of bacterial sRNAs, it is generally believed that the currently identified sRNAs constitute a limited subset of the bacterial sRNA repertoire. In several cases, sRNAs belonging to a specific class are already known and the challenge is to identify additional sRNAs belonging to the same class. **InvenireSRNA** is an R package for learning a classification model for a given class of sRNA thus allowing for the discovery of additional sRNAs beloning to the same class. **InvenireSRNA** also provides a pretrained model for predicting RsmA/CsrA regulating sRNAs.
}
\details{
\tabular{ll}{
Package: \tab InvenireSRNA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-08-01\cr
License: \tab GPL (>= 3)\cr
}

}
\author{
Carl Tony Fakhry

Maintainer: Carl Tony Fakhry <cfakhry@cs.umb.edu>
}
\references{
Carl Tony Fakhry, Prajna Kulkarni, Ping Chen, Rahul Kulkarni and Kourosh Zarringhalam (2017). "Prediction of bacterial small RNAs in the RsmA (CsrA) and ToxT pathways: a machine learning approach." BMC Genomics, 18.
}
