% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{predict_fasta}
\alias{predict_fasta}
\title{Given a model for prediction, compute the probabilities of the new sequences.}
\usage{
predict_fasta(fasta_file, InvenireSRNA_model = NULL)
}
\arguments{
\item{fasta_file}{A path to a FASTA file containing the RNA sequences constructed with
nucleotides A,C,G,T and U.}

\item{InvenireSRNA_model}{An object with S3 class \emph{InvenireSRNA} to be used for prediction. If \emph{InvenireSRNA_model = NULL}
then the pretrained model to predict CsrA regulating sRNAs is used.}
}
\value{
A data frame containing the mean and standard deviation of the prediction
              probabilities using the given \emph{InvenireSRNA_model} object.
}
\description{
Given a model for prediction, compute the probabilities of the
             new sequences.
}
\author{
Carl Tony Fakhry , Kourosh Zarringhalam, Rahul Kulkarni and Ping Chen.
}
