% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{sRNA_prob}
\alias{sRNA_prob}
\title{Computes the probability of a sequence beloning to a certain class of sRNA.}
\usage{
sRNA_prob(rna_sequence, InvenireSRNA_model = NULL)
}
\arguments{
\item{rna_sequence}{RNA Sequence constructed with nucleotides A,C,G,T and U.}

\item{InvenireSRNA_model}{object of class InvenireSRNA to be used for prediction.
If \emph{InvenireSRNA_model = NULL} then the pretrained
model to predict CsrA regulating sRNAs is used.}
}
\value{
The functions returns a list containing the following elements:
\itemize{
  \item{\strong{probability} }{probability of the given sequence belonging to a given class of sRNA.}
  \item{\strong{std_deviation} }{the standard deviation of the prediction.}
}
}
\description{
This function computes the probability of a given sequence belonging
             to a certain class of sRNA.
}
