% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{TrainModel}
\alias{TrainModel}
\title{This function trains a model given a fixed training data set.}
\usage{
TrainModel(Y, X, bootstrap_iterations = 1000, Kfold = 10)
}
\arguments{
\item{Y}{Class labels of the rows of \emph{X}. There can only be 2 class labels}

\item{X}{The feature matrix.}

\item{bootstrap_iterations}{The number of iterations for bootstrapping.}

\item{Kfold}{An integer specifying \emph{K} fold cross-validation.}
}
\value{
return the coefficients for the learned model. Non-robust features
        will be set to zero.
}
\description{
Given features for a class of sRNA, this function will learn robust
             features for the given class of sRNA and return the coefficients
             for all the features. Non-robust features will have coefficients
             equal to zero.
}
\author{
Carl Tony Fakhry , Kourosh Zarringhalam, Rahul Kulkarni and Ping Chen.
}
