% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{TrainModels}
\alias{TrainModels}
\title{This function trains a class of models given a FASTA file with seed sequences.}
\usage{
TrainModels(seed_sequences_fasta, n_models = 100, bootstrap_iterations = 1000)
}
\arguments{
\item{seed_sequences_fasta}{The FASTA file containing the seed sequences.}

\item{n_models}{The number of models to be learned. \emph{n_models} must be
greater than zero.}

\item{bootstrap_iterations}{The number of iterations for bootstrapping.}
}
\value{
an object of class InvenireSRNA which has a data frame \emph{coefficients_4_features}
        containing the coefficients of the Boltzmann Triplet features for
        \emph{bootstrap_iterations} number of models trained for prediction.

An object with S3 class \emph{InvenireSRNA} which contains the following elements:
\itemize{
  \item{\strong{coefficients_4_features} }{a data frame containing the coefficients of the Boltzmann Triplet features for
        \emph{bootstrap_iterations} number of models trained for prediction.}
}
}
\description{
Given a FASTA file of seed sequences of sRNA, we learn a set of models
             for prediction on the class of sRNAs which the seed sequences belong
             to.
}
\author{
Carl Tony Fakhry , Kourosh Zarringhalam, Rahul Kulkarni and Ping Chen.
}
